# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ChooseWorkshopRequest": {
        "properties": {
          "final_prompt": {
            "default": "",
            "title": "Final Prompt",
            "type": "string"
          },
          "option_id": {
            "title": "Option Id",
            "type": "string"
          }
        },
        "required": [
          "option_id"
        ],
        "title": "ChooseWorkshopRequest",
        "type": "object"
      },
      "ChosenPrompt": {
        "properties": {
          "prompt": {
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "ChosenPrompt",
        "type": "object"
      },
...
      "CreateWorkshopRequest": {
        "properties": {
          "context_docs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WorkshopContextDoc"
            },
            "title": "Context Docs",
            "type": "array"
          },
          "harness": {
            "$ref": "#/components/schemas/Harness",
            "default": "claude"
          },
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "CreateWorkshopRequest",
        "type": "object"
      },
...
      "GroundingDepth": {
        "enum": [
          "shallow",
          "standard",
          "deep"
        ],
        "title": "GroundingDepth",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Harness": {
        "enum": [
          "claude",
          "codex",
          "demo"
        ],
        "title": "Harness",
        "type": "string"
      },
...
      "RefineWorkshopRequest": {
        "properties": {
          "base": {
            "description": "An option id, or 'merge' to combine both.",
            "title": "Base",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "base"
        ],
        "title": "RefineWorkshopRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "ctx",
          "input",
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Workshop": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkshopError"
              },
              {
                "type": "null"
              }
            ]
          },
          "harness": {
            "$ref": "#/components/schemas/Harness"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "options": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WorkshopOption"
            },
            "title": "Options",
            "type": "array"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/WorkshopState"
          }
        },
        "required": [
          "created_at",
          "error",
          "harness",
          "id",
          "options",
          "question",
          "state"
        ],
        "title": "Workshop",
        "type": "object"
      },
      "WorkshopContextDoc": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "name",
          "text"
        ],
        "title": "WorkshopContextDoc",
        "type": "object"
      },
      "WorkshopError": {
        "description": "Why a workshop stopped. `code` is for the app, `message` is for the scientist.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail",
          "message"
        ],
        "title": "WorkshopError",
        "type": "object"
      },
      "WorkshopOption": {
        "properties": {
          "chosen": {
            "default": false,
            "title": "Chosen",
            "type": "boolean"
          },
          "excludes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excludes"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "optimizes_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimizes For"
          },
          "ordinal": {
            "title": "Ordinal",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "recommended_settings": {
            "$ref": "#/components/schemas/WorkshopRecommendedSettings"
          },
          "rejected": {
            "default": false,
            "title": "Rejected",
            "type": "boolean"
          },
          "strategy": {
            "title": "Strategy",
            "type": "string"
          }
        },
        "required": [
          "chosen",
          "excludes",
          "id",
          "note",
          "optimizes_for",
          "ordinal",
          "prompt",
          "rationale",
          "recommended_settings",
          "rejected",
          "strategy"
        ],
        "title": "WorkshopOption",
        "type": "object"
      },
      "WorkshopRecommendedSettings": {
        "description": "What this option wants the run configured as. Defaults match the Standard preset.",
        "properties": {
          "budget_calls": {
            "default": 60,
            "title": "Budget Calls",
            "type": "integer"
          },
          "grounding_depth": {
            "$ref": "#/components/schemas/GroundingDepth",
            "default": "standard"
          },
          "matches_per_round": {
            "default": 6,
            "title": "Matches Per Round",
            "type": "integer"
          },
          "rounds": {
            "default": 3,
            "title": "Rounds",
            "type": "integer"
          }
        },
        "required": [
          "budget_calls",
          "grounding_depth",
          "matches_per_round",
          "rounds"
        ],
        "title": "WorkshopRecommendedSettings",
        "type": "object"
      },
      "WorkshopState": {
        "enum": [
          "refining",
          "options_ready",
          "chosen",
          "failed"
        ],
        "title": "WorkshopState",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Local-first API for Oracle run orchestration and artifact discovery.",
    "title": "Oracle GUI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workshops": {
      "post": {
        "description": "Open a workshop. Returns in state `refining`; poll `GET` for the options.",
        "operationId": "create_workshop_api_workshops_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkshopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workshop"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workshop",
        "tags": [
          "workshops"
        ]
      }
    },
    "/api/workshops/{workshop_id}": {
      "get": {
        "description": "The workshop as it stands. Idempotent — this endpoint never writes.",
        "operationId": "get_workshop_api_workshops__workshop_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workshop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workshop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workshop"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workshop",
        "tags": [
          "workshops"
        ]
      }
    },
    "/api/workshops/{workshop_id}/choose": {
      "post": {
        "description": "Settle on an option. The scientist's edited prompt wins, stored exactly as typed.",
        "operationId": "choose_workshop_option_api_workshops__workshop_id__choose_post",
        "parameters": [
          {
            "in": "path",
            "name": "workshop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workshop Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChooseWorkshopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChosenPrompt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Choose Workshop Option",
        "tags": [
          "workshops"
        ]
      }
    },
    "/api/workshops/{workshop_id}/refine": {
      "post": {
        "description": "Reject the current pair and ask again, building on one option or merging both.",
        "operationId": "refine_workshop_api_workshops__workshop_id__refine_post",
        "parameters": [
          {
            "in": "path",
            "name": "workshop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workshop Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineWorkshopRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workshop"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refine Workshop",
        "tags": [
          "workshops"
        ]
      }
    },
...
  }
}